ecommerce-backend/requirements.txt
asgiref==3.3.1
boto3==1.17.22
botocore==1.20.22
Django==3.0.5
django-cors-headers==3.7.0
django-multiselectfield==0.1.12
django-rest-multiple-models==2.1.3
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangotoolbox==1.8.0
jmespath==0.10.0
jsonfield==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.3
